+++ onboarding-enabler-java/src/test/java/org/zowe/apiml/eurekaservice/client/util/EurekaInstanceConfigCreatorTest.java
@@ -13,12 +13,15 @@ package org.zowe.apiml.eurekaservice.client.util;
 import com.netflix.appinfo.EurekaInstanceConfig;
 import org.junit.Test;
 import org.zowe.apiml.eurekaservice.client.config.ApiMediationServiceConfig;
+import org.zowe.apiml.exception.MetadataValidationException;
 import org.zowe.apiml.exception.ServiceDefinitionException;
 
 import java.net.MalformedURLException;
 
 import static org.hamcrest.Matchers.hasEntry;
 import static org.junit.Assert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 public class EurekaInstanceConfigCreatorTest {
 
@@ -38,5 +41,13 @@ public class EurekaInstanceConfigCreatorTest {
         assertThat(translatedConfig.getMetadataMap(), hasEntry("customService.evenmorelevels.key5.key6.key7", "value7"));
     }
 
+    @org.junit.jupiter.api.Test
+    public void givenConfigurationWithInvalidProtocol_whenValidate_thenThrowException() throws ServiceDefinitionException {
+        ApiMediationServiceConfig testConfig = configReader.loadConfiguration("bad-protocol-baseurl-service-configuration.yml");
+        Exception exception = assertThrows(MetadataValidationException.class,
+            () -> eurekaInstanceConfigCreator.createEurekaInstanceConfig(testConfig),
+            "Expected exception is not MetadataValidationException");
+        assertEquals("'ftp' is not valid protocol for baseUrl property", exception.getMessage());
+    }
 
 }
